[MM-59349]: Getting Rid of Mattermost Auth layer (#10)
* refactor: get rid of mattermost auth layer

* chore: removed removeDuplicate from isboardsFromRow

* refactor: changed method to private'

* refactor: private method calls private method

* refactor: get license from server API
