+++ server/boards/boardsapp.go
@@ -14,7 +14,6 @@ import (
 	"github.com/mattermost/mattermost-plugin-boards/server/services/notify"
 	"github.com/mattermost/mattermost-plugin-boards/server/services/permissions/mmpermissions"
 	"github.com/mattermost/mattermost-plugin-boards/server/services/store"
-	"github.com/mattermost/mattermost-plugin-boards/server/services/store/mattermostauthlayer"
 	"github.com/mattermost/mattermost-plugin-boards/server/services/store/sqlstore"
 	"github.com/mattermost/mattermost-plugin-boards/server/ws"
 	"github.com/mattermost/mattermost/server/public/pluginapi/cluster"
@@ -90,11 +89,6 @@ func NewBoardsApp(api model.ServicesAPI) (*BoardsApp, error) {
 	if err != nil {
 		return nil, fmt.Errorf("error initializing the DB: %w", err)
 	}
-	layeredStore, err2 := mattermostauthlayer.New(cfg.DBType, sqlDB, db, logger, api, storeParams.TablePrefix)
-	if err2 != nil {
-		return nil, fmt.Errorf("error initializing the DB: %w", err2)
-	}
-	db = layeredStore
 
 	permissionsService := mmpermissions.New(db, api, logger)
