+++ server/services/store/generators/transactional_store.go.tmpl
@@ -16,7 +16,7 @@ import (
 	"context"
 	"time"
 
-    "github.com/mattermost/focalboard/server/model"
+    "github.com/mattermost/mattermost-plugin-boards/server/model"
 
     "github.com/mattermost/mattermost/server/public/shared/mlog"
 	mmModel "github.com/mattermost/mattermost/server/public/model"
